Move console.ts to utils/ and utils/ up one dir level
These utilities can be shared between all bin scripts.
