+++ src/bin/utils/console.ts
@@ -1,7 +1,7 @@
 import pinoPretty from 'pino-pretty'
 import { createInterface } from 'readline'
 
-import { print, statusBarLine, statusBox } from './utils/printing'
+import { print, statusBarLine, statusBox } from './printing'
 
 const pretty = pinoPretty({})
 
@@ -12,7 +12,7 @@ const rl = createInterface({
 rl.on('line', line => {
   const log = JSON.parse(line)
 
-  if (log.msg === 'status') {
+  if (log.msg === 'console') {
     statusBox(log.status.join('\n' + statusBarLine() + '\n'))
   } else {
     print(pretty(log))
