+++ src/bin/utils/metric.ts
